// src/apis/models/RiskClassificationScore.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RiskLevel } from './RiskLevel';
import { HttpFile } from '../http/http';

export class RiskClassificationScore {
  'riskLevel': RiskLevel;
  'lowerBoundRiskScore': number;
  'upperBoundRiskScore': number;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'riskLevel',
      baseName: 'riskLevel',
      type: 'RiskLevel',
      format: '',
    },
    {
      name: 'lowerBoundRiskScore',
      baseName: 'lowerBoundRiskScore',
      type: 'number',
      format: '',
    },
    {
      name: 'upperBoundRiskScore',
      baseName: 'upperBoundRiskScore',
      type: 'number',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return RiskClassificationScore.attributeTypeMap;
  }

  public constructor() {}
}